backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
torchvision
opencv-python-headless
numpy
pydantic
scikit-learn
matplotlib
seaborn